requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.1.0
black==23.9.1
blinker==1.6.2
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
decorator==5.1.1
Flask==2.3.3
frozenlist==1.4.0
idna==3.4
importlib-metadata==6.8.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
multidict==6.0.4
mypy-extensions==1.0.0
openai==0.28.0
packaging==23.1
pathspec==0.11.2
platformdirs==3.10.0
py==1.11.0
python-decouple==3.8
requests==2.31.0
retry==0.9.2
tomli==2.0.1
tqdm==4.66.1
typing_extensions==4.7.1
urllib3==2.0.4
Werkzeug==2.3.7
yarl==1.9.2
zipp==3.16.2
